+++ src/Planner.Contracts/Optimization/Outputs/TaskAssignment.cs
@@ -4,7 +4,7 @@
 /// Describes one stop in an optimized vehicle route.
 /// </summary>
 public sealed record TaskAssignment(
-    int JobId,
+    long JobId,
     int JobType,
     string Name,
     double ArrivalTime,
